Refactor constants for option lab components
- Consolidated option lab constants into a single file for better organization and maintainability.
- Updated imports in collateral-provider, option-lab-form, place-trade-order, and calculations to reflect the new constants structure.
- Removed the deprecated option-lab constants file to streamline the codebase.
